Move python file in different folder. Create ask_metadata. Use Llama3/groq API
