--- python_workspace/pyproject.toml
@@ -1,35 +1,0 @@
-[tool.poetry]
-name = "forestbot workspace"
-version = "0.1.0"
-description = ""
-authors = ["Guillaume Soto", "Tony Lopes", "Connor Haines"]
-package-mode = false
-
-[tool.poetry.dependencies]
-python = "^3.10,<3.11"
-openai = "^1.17.1"
-langchain = { extras = ["docarray"], version = "^0.1.16" }
-gunicorn = "^22.0.0"
-langchain-openai = "^0.1.3"
-python-dotenv = "^1.0.1"
-panel = "^1.4.1"
-param = "^2.1.0"
-ipykernel = "^6.29.4"
-pypdf = "^4.2.0"
-streamlit = "^1.33.0"
-langchain-community = "^0.0.33"
-chromadb = "^0.4.24"
-streamlit-chat = "^0.1.1"
-icecream = "^2.1.3"
-pyzmq = "26.0.0"
-ragas = "^0.1.7"
-matplotlib = "^3.8.4"
-
-# this is a workaround for the pysqlite3-binary package not being available for Chroma version in Streamlit deployment
-# pysqlite3-binary = { version = "*", markers="sys_platform == 'darwin' and platform_machine == 'arm64'" }
-pysqlite3-binary = { version = "*", markers = "platform_machine == 'x86_64'" }
-
-
-[build-system]
-requires = ["poetry-core"]
-build-backend = "poetry.core.masonry.api"
